Disable client tools auto update disabled if there are no home dir Move updater to general tools package
